fer_genai/requirements_updated.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
bcrypt==4.0.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.1
filelock==3.24.3
Flask==3.1.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.64.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lightning-utilities==0.15.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.4.2
numpy==2.0.2
opencv-python==4.12.0.88
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pillow==11.3.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
timm==1.0.24
torch==2.9.0
torchmetrics==1.8.2
torchvision==0.24.0
tqdm==4.67.3
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
Werkzeug==3.1.6
yarl==1.22.0
